Finished refactoring Amount,

+++ src/main/java/nl/gogognome/gogoaccount/gui/dialogs/JournalEntryDetailsTableModel.java
@@ -1,17 +1,13 @@
 package nl.gogognome.gogoaccount.gui.dialogs;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
 import nl.gogognome.gogoaccount.component.configuration.Account;
 import nl.gogognome.gogoaccount.component.configuration.ConfigurationService;
+import nl.gogognome.gogoaccount.component.document.Document;
 import nl.gogognome.gogoaccount.component.invoice.Invoice;
-import nl.gogognome.gogoaccount.component.ledger.JournalEntryDetail;
 import nl.gogognome.gogoaccount.component.invoice.InvoiceService;
+import nl.gogognome.gogoaccount.component.ledger.JournalEntryDetail;
 import nl.gogognome.gogoaccount.component.party.Party;
 import nl.gogognome.gogoaccount.component.party.PartyService;
-import nl.gogognome.gogoaccount.component.document.Document;
 import nl.gogognome.gogoaccount.services.ServiceException;
 import nl.gogognome.gogoaccount.util.ObjectFactory;
 import nl.gogognome.lib.swing.AbstractListTableModel;
@@ -21,6 +17,10 @@ import nl.gogognome.lib.text.AmountFormat;
 import nl.gogognome.lib.util.Factory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
 
 public class JournalEntryDetailsTableModel extends AbstractListTableModel<JournalEntryDetail> {
 
@@ -79,9 +79,9 @@ public class JournalEntryDetailsTableModel extends AbstractListTableModel<Journa
                 result = "???";
             }
         } else if (DEBET == colDef) {
-            result = journalEntryDetail.isDebet() ? af.formatAmountWithoutCurrency(journalEntryDetail.getAmount()) : "" ;
+            result = journalEntryDetail.isDebet() ? af.formatAmountWithoutCurrency(journalEntryDetail.getAmount().toBigInteger()) : "" ;
         } else if (CREDIT == colDef) {
-            result = journalEntryDetail.isCredit() ? af.formatAmountWithoutCurrency(journalEntryDetail.getAmount()) : "" ;
+            result = journalEntryDetail.isCredit() ? af.formatAmountWithoutCurrency(journalEntryDetail.getAmount().toBigInteger()) : "" ;
         } else if (INVOICE == colDef) {
             if (journalEntryDetail.getInvoiceId() != null) {
                 try {
